Implemented Factory pattern
